### 2024-03-25 #170 @sanjeevrs2000
CovPlan
Submitting Author: Name (@sanjeevrs2000)
Package Name: CovPlan
One-Line Description of Package: A Python package that generates guidance trajectories for field coverage using a single robot.
Repository Link:  https://github.com/sanjeevrs2000/covplan
Version submitted:  0.1.0
Version accepted: TBD

---

This Python package generates a guidance trajectory for complete coverage in 2 dimensions. It can be used for operations where complete coverage of an Area of Interest (AoI) is required for applications in field robotics. If the area of interest is large or if it has any forbidden regions or obstacles, it could be divided into smaller sections and covered sequentially where the sequence is optimized using a travelling salesman problem (TSP) solver to minimize the overall distance. It also uses Dubins curves to generate continuous and feasible trajectories.

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.

It can be classified under applications in AI and robotics. Unsure which of the above categories it comes under but were encouraged to make a submission in the [presubmission enquiry](https://github.com/pyOpenSci/software-submission/issues/140). Researchers in coverage path planning might find it useful for developing new algorithms or comparing it against their own methods. It has potential applications in field robotics where coverage of an area is required for monitoring, information gathering tasks.

[Fields2Cover](https://github.com/Fields2Cover/Fields2Cover) is a similar implementation although it is in C++. Our work also considers more complex areas for coverage such as ones with forbidden regions and solves it by using a divide and conquer approach. It is also not limited to agriculture related applications and can be adapted for other uses by modifying the parameters

The presubmission enquiry can be found [here](https://github.com/pyOpenSci/software-submission/issues/140).
