### 2024-04-09 #170 @Batalex
Hello @sanjeevrs2000,
Thank you for your patience. You may not know, but we now implement a 3-month rotation for the editor-in-chief role. I'll get started with the EiC checks right now.
