### 2024-04-09 #170 @Batalex
Hi @Gonzalo-Mier, welcome to pyOpenSci!

Thanks for the info. As @NickleDave reminded me, Fields2Cover was mentioned in the [presubmission inquiry](https://github.com/pyOpenSci/software-submission/issues/140#issuecomment-1758622411) a few months ago.
As we already decided that `CovPlan` was in scope, we most likely will proceed with the review.

Now for the shameless part 😁
Though we do try to avoid duplication of efforts (hence our question on similar packages), we do not position ourselves as an authority on which package should be the reference. If anything, I would love for this GH issue to be the starting point of a collaborative effort that would benefit everyone.
If you are interested and available, please let me know!
