### 2024-04-09 #170 @Batalex
Hi there! Thank you for submitting your package for pyOpenSci review. Below are the basic checks that your package needs to pass to begin our review. If some of these are missing, we will ask you to work on them before the review process begins.

  ✨ **Please add some docstrings to document user-facing functions.**
  ✨   **I think we could use some more structure in the readme. "Getting started" should be like a 30 sec tutorial, not the description and even less the reference paper. See below for a suggestion**

---
Thank you in advance for setting aside five to ten minutes to do this. It truly helps our organization. :raised_hands:
---

As of now, there are too many missing things to get started with the review. Some are really trivial (like adding a `CONTRIBUTING.md` or restructuring the readme using the suggestion below), others require additional work.

I encourage you to go through our [Python Package Guide](https://www.pyopensci.org/python-package-guide/) to get familiar with:
- [Python docstrings](https://peps.python.org/pep-0257/), and why you should use them to document your code
- Unit testing to avoid regressions in your code
- How to structure your package files

### README.md suggested structure

<details><summary>Expand</summary>

````md

# CovPlan

A Python package for coverage path planning.

A brief description of what this project does and who it's for.

Link to paper + full citation

> Hameed IA, Bochtis D, Sørensen CA. An Optimized Field Coverage Planning Approach for Navigation of Agricultural Robots in Fields Involving Obstacle Areas. International Journal of Advanced Robotic Systems. 2013;10(5). doi:10.5772/56248

Install `covplan` with pip

pip install covplan
```

## Getting started

The 30 seconds tutorial.

```py
import antigravity
```

[Documentation](https://linktodocumentation)

## Contributing

Contributions are always welcome!

See `contributing.md` for ways to get started.

Please adhere to this project's `code of conduct`.

````
