### 2024-04-12 #170 @sanjeevrs2000
Hi @Gonzalo-Mier, thanks again for your support. I am not a software developer myself so I have limited time as well. Will be happy to collaborate if required.

> I don't have too much free time, but I can help if needed. Please, contact me for any collaboration :)

@Batalex, if it is decided that it is under scope for a review, we shall try to meet these requirements as soon as possible.

> Some are really trivial (like adding a `CONTRIBUTING.md` or restructuring the readme using the suggestion below), others require additional work
