+++ README.md
@@ -6,10 +6,16 @@ A Python package for coverage path planning
 [![PyPi license](https://img.shields.io/pypi/l/ansicolortags.svg)](https://pypi.org/project/covplan/0.1.0/)
 
 This repository can be used for generating guidance trajectories for complete field coverage. It can be used for operations where complete coverage of an Area of Interest (AoI) is required for various applications. It is an updated Python implementation of the method that was presented in [this paper](https://journals.sagepub.com/doi/full/10.5772/56248).
+> Hameed IA, Bochtis D, Sørensen CA. An Optimized Field Coverage Planning Approach for Navigation of Agricultural Robots in Fields Involving Obstacle Areas. International Journal of Advanced Robotic Systems. 2013;10(5). doi:10.5772/56248
+
 
 ## Installation
 
-The package can be installed from PyPi by running `pip install covplan`. 
+The package can be installed using pip 
+
+```shell
+pip install covplan
+```
 
 ## Getting started
 
