+++ contributing.md
@@ -1,0 +1,49 @@
+# Contributing to CovPlan
+Thank you for your interest in contributing to CovPlan! We welcome contributions from everyone. By contributing to this project, you agree to abide by our Code of Conduct.
+
+## How to Contribute
+
+### Reporting Bugs
+If you encounter any bugs, please open an issue on our issue tracker. Please provide a clear description of the problem, steps to reproduce it, and any relevant error messages or screenshots.
+
+## Suggesting Enhancements
+If you have an idea for a new feature or an enhancement, please open an issue on our issue tracker and describe your suggestion in detail.
+
+### Contributing Code
+Fork the repository and create a new branch for your contribution.
+Write tests for your changes and ensure that all tests pass.
+Update the documentation if necessary.
+Commit your changes with a descriptive commit message.
+Push your changes to your forked repository.
+Open a pull request against the main repository, and provide a clear description of your changes.
+
+## License
+By contributing to this project, you agree that your contributions will be licensed under the project's [license](/LICENSE).
+
+## Thank You!
+Thank you for your interest in contributing to CovPlan! We appreciate your efforts to improve the project.
+
+
+## Code of Conduct
+Please note that this project is released with a Code of Conduct. By participating in this project, you agree to abide by its terms.
+
+### Our Pledge
+We as members, contributors, and leaders pledge to make participation in our community a harassment-free experience for everyone.
+
+### Our Standards
+Examples of behavior that contributes to a positive environment include:
+Being respectful of differing opinions and experiences
+Gracefully accepting constructive feedback
+Focusing on what is best for the community
+
+Unacceptable behavior includes:
+Personal attacks
+Trolling or insulting/derogatory comments
+Public or private harassment
+
+### Enforcement
+
+Community leaders have the right and responsibility to remove, edit, or reject comments, commits, code, wiki edits, issues, and other contributions that are not aligned with this Code of Conduct.
+
+### Attribution
+This Code of Conduct is adapted from the Contributor Covenant, version 2.1, available at https://www.contributor-covenant.org/version/2/1/code_of_conduct.html.
