+++ README.md
@@ -56,3 +56,11 @@ if __name__ == '__main__':
 ## Documentation
 
 The full documentation can be found [here](https://covplan.readthedocs.io/).
+
+## Contributing
+
+Contributions are always welcome!
+
+See [contributing.md](/contributing.md) for ways to get started.
+
+Please adhere to this project's [code of conduct](/contributing.md#code-of-conduct)
