+++ build/lib/covplan/__init__.py
@@ -1,0 +1,2 @@
+from covplan import *
+from .coverage_path_planner import find_min,pathplan
