+++ src/covplan/sample_area.txt
@@ -1,0 +1,11 @@
+53.0    1.0
+53.0    1.002
+52.999   1.002
+52.999   1.0
+53.0    1.0
+NaN NaN
+52.9994    1.0015
+52.9996    1.001
+52.9997    1.0015
+52.9994    1.0015
+NaN NaN
