+++ docs/index.md
@@ -1,8 +1,8 @@
 # Introduction
 
-This Python package generates a guidance trajectory for complete coverage in 2 dimensions. It can be used for operations where complete coverage of an Area of Interest (AoI) is required for applications in field robotics. This can also assist researchers in coverage path planning who want to compare it their algorithms  or want to develop new methods . It is an updated Python implementation of the coverage method discussed in [this paper](https://journals.sagepub.com/doi/full/10.5772/56248).
+This Python package generates a guidance trajectory for complete coverage in 2 dimensions. It can be used for operations where complete coverage of an Area of Interest (AoI) is required for applications in field robotics. It can also assist researchers in coverage path planning who want to compare it their algorithms or want to develop new methods. It is an updated Python implementation of the coverage method discussed in [this paper](https://journals.sagepub.com/doi/full/10.5772/56248).
 
- It takes in only coordinates of points on the boundary of the area in consideration and returns the full trajectory as a list of latitude-longitude coordinates which makes it easily transferrable to potential applications. 
+It uses an area-based approach as opposed to a grid-map based approach that is used in most studies. It takes in only coordinates of points describing boundaries of the AoI in consideration and returns the full path as a list of latitude-longitude coordinates, that makes it easily transferrable to real-life applications. 
 
 # Getting Started
 
