### 2024-05-21 #170 @sanjeevrs2000
Hi @Batalex,
I've added tests and updated the documentation as per the recommendations. Hopefully, this is enough to get started with the review now. Let me know if you find any other areas for improvement.
