### 2024-05-25 #170 @Batalex
Hi @sanjeevrs2000,
Thank you for your work on the package. There are still some things missing before I can give the go for the review.

## Documentation

Since you are using mkdocs, I advise you to take a look at https://mkdocstrings.github.io/. It is a plugin that automatically picks up the docstrings to embed them in your documentation. The reason why I ask you to do that is because it will help you keep your code base and your documentation in sync.
Right now, [the API reference](https://covplan.readthedocs.io/en/latest/usage/) looks like it's been written manually, and it is a problem. [See](https://github.com/sanjeevrs2000/covplan/blob/main/src/covplan/coverage_path_planner.py#L6-L16), the code base is already out of sync with the documentation, and this issue will only continue to grow with the code base and the number of contributors to the project.

I also noticed a lack of consistency in the docstrings formats: some use Numpy/Google convention, others use the classic Python one I linked in my previous post. The function linked above uses neither.

You might want to explicitly list the development dependencies one needs to work on the project. It happens that I am quite familiar with mkdocs so I already know how to build the documentations, but that most likely not the case for everyone.
To do so, you may:
- add mkdocs (and mkdocstrings 👼 ) to the `pyproject.toml` file using the [optional deps field ](https://packaging.python.org/en/latest/guides/writing-pyproject-toml/#dependencies-and-requirements)(since you are using the setuptools build backend).
- document somewhere how to build the docs (contributing.md)

## Tests and automation

I am glad to see that you added tests, but we need to go one step further. Because tests are only as useful as their consistent runs, we need to integrate them into GitHub.
Even before that, same as the previous section, it is a good idea to include a mention of how to run the tests in your contributing.md file. Same as above, you will need to explicitly state your test dependencies if you have any.

Back to GitHub, the aim of the continuous integration (CI), is to make sure that you do not introduce any regression to the code base, using the tests you wrote. By making GitHub run your tests on every pull request (and I advise you to use this workflow from now on even if you are the sole contributor), you can merge new additions to the code base more confidently.

We wrote [a guide](https://www.pyopensci.org/python-package-guide/tests/tests-ci.html) for that, this is a great starting point.

They are a few more things we could work on, but I don't want to move the goal posts further with every interaction, so I'll leave it there for now.
